+++ src/Search/GridHelperFunctions.js
